Fix imports, consolidate some checkpoint code
